requirements.txt
asana
openai
python-dotenv
requests